+++ packages/miniapp-render/src/builtInComponents/textarea.js
@@ -1,7 +1,11 @@
-export default {
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { isWeChatMiniProgram, isBaiduSmartProgram } from 'universal-env';
+import { isUndef } from '../utils/tool';
+
+const textarea = {
   name: 'textarea',
   singleEvents: [{
-    name: 'keyboardheightchange',
+    name: 'onTextareaKeyboardHeightChange',
     eventName: 'keyboardheightchange'
   }],
   simpleEvents: [{
@@ -22,7 +26,7 @@ export default {
     eventName: 'blur',
     middleware(evt, domNode, nodeId) {
       domNode._setAttributeWithOutUpdate('focus-state', false);
-      if (domNode.__textareaOldValue !== undefined && domNode.value !== domNode.__textareaOldValue) {
+      if (!isUndef(domNode.__textareaOldValue) && domNode.value !== domNode.__textareaOldValue) {
         domNode.__textareaOldValue = undefined;
         this.callEvent('change', evt, null, nodeId);
       }
@@ -40,3 +44,12 @@ export default {
     }
   }]
 };
+
+if (isWeChatMiniProgram || isBaiduSmartProgram) {
+  textarea.simpleEvents = textarea.simpleEvents.concat([{
+    name: 'onTextareaLineChange',
+    eventName: 'linechange'
+  }]);
+}
+
+export default textarea;