+++ packages/miniapp-render/src/constants.js
@@ -23,3 +23,13 @@ export const BUILTIN_COMPONENT_LIST = new Set([
 ]);
 
 export const BODY_NODE_ID = 'e-body';
+
+export const STATIC_COMPONENTS = new Set(['view', 'text', 'image']); // With no events components
+
+export const PURE_COMPONENTS = new Set(['view', 'h-element']); // With no events or props && equal to TOUCH_COMPONENTS
+
+export const CATCH_COMPONENTS = new Set(['view', 'h-element']); // With catchTouchMove events
+
+export const APPEAR_COMPONENT = 'view'; // Without appear event components
+
+export const ANCHOR_COMPONENT = 'scroll-view'; // Components which only use scrollIntoView to scroll