+++ packages/miniapp-render/src/node/node.js
@@ -1,5 +1,7 @@
+// eslint-disable-next-line import/no-extraneous-dependencies
+import { isBaiduSmartProgram } from 'universal-env';
 import EventTarget from '../event/event-target';
-import tool from '../utils/tool';
+import { getId } from '../utils/tool';
 import cache from '../utils/cache';
 import { BODY_NODE_ID } from '../constants';
 
@@ -8,9 +10,10 @@ class Node extends EventTarget {
     super();
 
     // unique node id
-    this.__nodeId = `n_${tool.getId()}`;
-    this.$_type = options.type;
+    this.__nodeId = `n_${getId()}`;
+    this.__type = options.type;
     this.parentNode = null;
+    this.__rendered = false;
     this.__ownerDocument = options.document;
   }
 
@@ -23,21 +26,21 @@ class Node extends EventTarget {
   }
 
   /**
-   * Override parent class $$destroy method
+   * Override parent class _destroy method
    */
-  $$destroy() {
-    super.$$destroy();
+  _destroy() {
+    super._destroy();
 
     this.__nodeId = null;
-    this.$_type = null;
+    this.__type = null;
     this.parentNode = null;
     this.__rendered = false;
-    this.__ownerDocument = null;
   }
 
   get _path() {
     if (this.parentNode !== null) {
-      const index = '[' + this.parentNode.childNodes.indexOf(this) + ']';
+      const childIndex = this.parentNode.childNodes.indexOf(this);
+      const index = isBaiduSmartProgram ? childIndex : `[${childIndex}]`;
 
       return `${this.parentNode._path}.children.${index}`;
     }