+++ packages/rax-miniapp-runtime-webpack-plugin/src/generators/templates/recursiveTemplate.js
@@ -10,11 +10,11 @@ const platformConfig = require('../../platforms');
 function buildRecursiveTemplate(target, customComponentsConfig) {
   const { internalComponents, derivedComponents, sjs, adapter } = platformConfig[target];
 
-  let template = buildBaseTemplate(sjs, {});
+  let template = buildBaseTemplate(sjs, { adapter });
   template += buildChildrenTemplate(0, adapter, {});
 
   const customInternalComponents = modifyInternalComponents(internalComponents, customComponentsConfig);
-  const miniComponents = createMiniComponents(customInternalComponents, adapter);
+  const miniComponents = createMiniComponents(customInternalComponents, derivedComponents, adapter);
   const components = Object.keys(miniComponents);
 
   template = components.reduce((current, nodeName) => {