+++ packages/rax-miniapp-runtime-webpack-plugin/src/platforms/ali.js
@@ -41,9 +41,37 @@ const View = {
 
 const CatchView = Object.assign({}, View);
 
-const HElement = {
+const StaticView = {
   props: {
+    'disable-scroll': 'false',
+    'hover-class': '',
+    'hover-start-time': '',
+    'hover-stay-time': '',
+    hidden: 'false',
+    'hover-stop-propagation': 'false',
+    role: '',
     animation: 'null',
+  },
+};
+
+const PureView = {};
+
+const NoAppearView = {
+  props: {
+    'disable-scroll': 'false',
+    'hover-class': '',
+    'hover-start-time': '',
+    'hover-stay-time': '',
+    hidden: 'false',
+    'hover-stop-propagation': 'false',
+    role: '',
+    animation: 'null',
+  },
+  events: {
+    TransitionEnd: '',
+    AnimationIteration: '',
+    AnimationStart: '',
+    AnimationEnd: '',
   },
   basicEvents: {
     ...tapEvents,
@@ -51,7 +79,73 @@ const HElement = {
   }
 };
 
+const NoTouchView = {
+  props: {
+    'disable-scroll': 'false',
+    'hover-class': '',
+    'hover-start-time': '',
+    'hover-stay-time': '',
+    hidden: 'false',
+    'hover-stop-propagation': 'false',
+    role: '',
+    animation: 'null',
+  },
+  events: {
+    TransitionEnd: '',
+    AnimationIteration: '',
+    AnimationStart: '',
+    AnimationEnd: '',
+    Appear: '',
+    Disappear: '',
+    FirstAppear: '',
+  },
+  basicEvents: {
+    ...tapEvents
+  }
+};
+
+const NoAppearTouchView = {
+  props: {
+    'disable-scroll': 'false',
+    'hover-class': '',
+    'hover-start-time': '',
+    'hover-stay-time': '',
+    hidden: 'false',
+    'hover-stop-propagation': 'false',
+    role: '',
+    animation: 'null',
+  },
+  events: {
+    TransitionEnd: '',
+    AnimationIteration: '',
+    AnimationStart: '',
+    AnimationEnd: '',
+  },
+  basicEvents: {
+    ...tapEvents
+  }
+};
+
+
+const HElement = {
+  props: {},
+  basicEvents: {
+    ...tapEvents,
+    ...touchEvents
+  }
+};
+
+const NoTouchHElement = {
+  props: {},
+  basicEvents: {
+    ...tapEvents
+  }
+};
+
+
 const CatchHElement = Object.assign({}, HElement);
+
+const PureHElement = {};
 
 const HComment = {};
 
@@ -179,6 +273,15 @@ const Text = {
   }
 };
 
+const StaticText = {
+  props: {
+    selectable: 'false',
+    space: '',
+    decode: 'false',
+    'number-of-lines': ''
+  }
+};
+
 const Icon = {
   props: {
     type: '',
@@ -226,6 +329,10 @@ const Button = {
     scope: '',
     'public-id': '',
     'data-params': '' // For share button
+  },
+  events: {
+    GetAuthorize: '',
+    Error: '',
   },
   basicEvents: {
     ...tapEvents
@@ -433,6 +540,15 @@ const Image = {
   }
 };
 
+const StaticImage = {
+  props: {
+    src: '',
+    mode: addSingleQuote('scaleToFill'),
+    'lazy-load': 'false',
+    'default-source': '',
+  }
+};
+
 const Video = {
   props: {
     src: '',
@@ -466,6 +582,25 @@ const Video = {
     UserAction: '',
     Stop: '',
     RenderStart: ''
+  }
+};
+
+const Camera = {
+  props: {
+    id: '',
+    mode: addSingleQuote('normal'),
+    'device-position': addSingleQuote('back'),
+    flash: addSingleQuote('auto'),
+    outputDimension: addSingleQuote('720P'),
+    applyMicPermissionWhenInit: 'true',
+    'frame-size': addSingleQuote('medium'),
+    'frame-format': addSingleQuote('rgba'),
+    'max-duration': '30'
+  },
+  events: {
+    Stop: '',
+    Error: '',
+    ScanCode: ''
   }
 };
 
@@ -611,6 +746,11 @@ const ContactButton = {
 exports.internalComponents = {
   View,
   CatchView,
+  StaticView,
+  PureView,
+  NoAppearView,
+  NoTouchView,
+  NoAppearTouchView,
   Swiper,
   SwiperItem,
   ScrollView,
@@ -619,6 +759,7 @@ exports.internalComponents = {
   MovableView,
   MovableArea,
   Text,
+  StaticText,
   Icon,
   Progress,
   RichText,
@@ -638,7 +779,9 @@ exports.internalComponents = {
   Picker,
   Navigator,
   Image,
+  StaticImage,
   Video,
+  Camera,
   Lottie,
   Canvas,
   Map: MiniappMap,
@@ -648,86 +791,100 @@ exports.internalComponents = {
   LivePusher,
   ContactButton,
   HElement,
+  NoTouchHElement,
   CatchHElement,
+  PureHElement,
   HComment
 };
 
 exports.derivedComponents = new Map([
-  ['catch-view', 'view'],
-  ['catch-h-element', 'view'],
-  ['h-element', 'view'],
-  ['h-comment', 'block']
+  ['CatchView', 'View'],
+  ['StaticView', 'View'],
+  ['PureView', 'View'],
+  ['NoAppearView', 'View'],
+  ['NoTouchView', 'View'],
+  ['NoAppearTouchView', 'View'],
+  ['CatchHElement', 'View'],
+  ['PureHElement', 'View'],
+  ['NoTouchHElement', 'View'],
+  ['HElement', 'View'],
+  ['StaticText', 'Text'],
+  ['StaticImage', 'Image'],
+  ['HComment', 'Block']
 ]);
 
 exports.controlledComponents = new Set([
-  'input',
-  'checkbox',
-  'picker',
-  'picker-view',
-  'radio',
-  'slider',
-  'switch',
-  'textarea'
+  'Input',
+  'Checkbox',
+  'Picker',
+  'PickerView',
+  'Radio',
+  'Slider',
+  'Switch',
+  'Textarea'
 ]);
 
 exports.focusComponents = new Set([
-  'input',
-  'textarea'
+  'Input',
+  'Textarea'
 ]);
 
 exports.voidElements = new Set([
-  'h-comment'
+  'HComment'
 ]);
 
 exports.voidChildrenElements = new Set([
-  'progress',
-  'icon',
-  'rich-text',
-  'input',
-  'textarea',
-  'slider',
-  'switch',
-  'live-pusher',
-  'h-comment',
-  'open-avatar',
-  'web-view',
-  'image',
-  'video',
-  'lottie',
-  'canvas',
-  'live-player',
-  'live-pusher',
-  'contact-button'
+  'Progress',
+  'Icon',
+  'RichText',
+  'Input',
+  'Textarea',
+  'Checkbox',
+  'Radio',
+  'Slider',
+  'Switch',
+  'LivePusher',
+  'HComment',
+  'Image',
+  'Video',
+  'Camera',
+  'Lottie',
+  'Canvas',
+  'WebView',
+  'LivePlayer',
+  'LivePusher',
+  'ContactButton',
+  'OpenAvatar'
 ]);
 
+
 exports.nestElements = new Map([
-  ['view', -1],
-  ['cover-view', -1],
-  // ['block', -1],
-  ['text', 6],
-  ['label', 6],
-  ['form', 4],
-  ['scroll-view', 4],
-  ['swiper', 4],
-  ['swiper-item', 4]
+  ['View', -1],
+  ['CoverView', -1],
+  ['Text', 6],
+  ['Label', 6],
+  ['Form', 4],
+  ['ScrollView', 4],
+  ['Swiper', 4],
+  ['SwiperItem', 4]
 ]);
 
 exports.shouldNotGenerateTemplateComponents = new Set([
-  'swiper-item',
-  'picker-view-column',
-  'movable-view'
+  'SwiperItem',
+  'PickerViewColumn',
+  'MovableView'
 ]);
 
 exports.needModifyChildrenComponents = {
-  swiper: children => `
+  Swiper: children => `
     <swiper-item a:for="{{r.children}}" a:key="nodeId">
       <template is="RAX_TMPL_CHILDREN_0" data="{{r: item.children}}" />
     </swiper-item>`,
-  'movable-area': children => `
+  MovableArea: children => `
     <movable-view a:for="{{r.children}}" a:key="nodeId" a:if="{{item.nodeType !== 'h-comment'}}" direction="{{item['direction']||'none'}}" inertia="{{tool.a(item['inertia'],false)}}" out-of-bounds="{{tool.a(item['out-of-bounds'],false)}}" x="{{tool.a(item['x'],0)}}" y="{{tool.a(item['y'],0)}}" damping="{{tool.a(item['damping'],20)}}" friction="{{tool.a(item['friction'],2)}}" disabled="{{tool.a(item['disabled'],false)}}" scale="{{tool.a(item['scale'],false)}}" scale-min="{{tool.a(item['scale-min'],0.5)}}" scale-max="{{tool.a(item['scale-max'],10)}}" scale-value="{{tool.a(item['scale-value'],1)}}" animation="{{tool.a(item['animation'],false)}}" onChange="onMovableViewChange" onChangeEnd="onMovableViewChangeEnd" onScale="onMovableViewScale" onTouchStart="onTouchStart" onTouchMove="onTouchMove" onTouchEnd="onTouchEnd" onTouchCancel="onTouchCancel" onLongTap="onLongTap" style="{{item.style}}" class="{{item.class}}" id="{{item.id}}" data-private-node-id="{{item.nodeId}}">
       <template is="RAX_TMPL_CHILDREN_0" data="{{r: item.children}}" />
     </movable-view>`,
-  'picker-view': children => `
+  PickerView: children => `
     <picker-view-column a:for="{{r.children}}" a:key="nodeId" a:if="{{item.nodeType !== 'h-comment'}}">
       <view a:for="{{item.children}}" a:for-item="pickerColumnItem">
         <block a:if="{{pickerColumnItem.nodeId}}">
@@ -738,7 +895,7 @@ exports.needModifyChildrenComponents = {
         </block>
       </view>
     </picker-view-column>`,
-  picker: children => `
+  Picker: children => `
     <view>
       ${children}
     </view>`
@@ -755,7 +912,9 @@ exports.adapter = {
   xs: 'sjs',
   event: 'on',
   catchEvent: 'catch',
-  eventToLowerCase: false
+  eventToLowerCase: false,
+  supportSjs: true,
+  formatBindedData: (value) => `${value}`
 };
 
 exports.sjs = {