+++ packages/jsx2mp-loader/CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## [0.4.28] - 2021-06-30
+
+### Fixed
+
+- Miss basedir config in resolve
+
+## [0.4.27] - 2021-06-29
+
+### Changed
+
+- Use resolve to replace `require.resolve` to make exports field compatible
+
+## [0.4.25] - 2021-05-19
+
+### Chore
+
+- Use constants in miniapp-builder-shared
+
+## [0.4.25] - 2021-03-30
+
+### Added
+
+- Support passing virtualHost config into jsx-compiler
 ## [0.4.24] - 2021-03-11
 
 ### Chore