+++ packages/miniapp-builder-shared/CHANGELOG.md
@@ -1,5 +1,35 @@
 # Changelog
 
+## [0.2.6] - 2021-06-01
+
+### Changed
+
+- Move transform `tabItem.path => tabItem.pagePath` logic to `rax-miniapp-config-webpack-plugin`
+
+## [0.2.5] - 2021-05-19
+
+### Changed
+
+- Update kuaishou miniprogram wxs extension name
+
+## [0.2.4] - 2021-05-10
+
+### Added
+
+- Add baidu and kuaishou config
+
+## [0.2.3] - 2021-04-27
+
+### Changed
+
+- Update homepage url in `package.json`
+
+## [0.2.2] - 2021-03-30
+
+### Added
+
+- `autoInstallNpm` method for installing npm for native miniapp automatically
+
 ## [0.2.1] - 2021-03-11
 
 ### Added