+++ packages/rax-miniapp-config-webpack-plugin/CHANGELOG.md
@@ -1,6 +1,25 @@
 # Changelog
 
-## [2.0.1] - 2021-03-11
+## [2.1.1] - 2021-06-01
+
+- Support transform tabBar config
+- Remove custom tabBar logic
+
+## [2.1.0] - 2021-05-19
+
+- Support bytedance microapp/baidu smartprogram/kuaishou miniprogram
+
+## [2.0.4] - 2021-05-10
+
+- Use constants from miniapp-builder-shared
+
+## [2.0.3] - 2021-03-31
+
+### Feat
+
+- Add `subPackageBuildType` when `shareMemory`
+
+## [2.0.2] - 2021-03-11
 
 ### Changed